chore: extract example from basic demo
