+++ double-ratchets/examples/double_ratchet_basic.rs
@@ -27,49 +27,4 @@ fn main() {
         "Alice received: {}",
         String::from_utf8_lossy(&plaintext.unwrap())
     );
-
-    // === Simulate alice and bob restarts
-    println!("Before restart, persist the state");
-    let alice_state = alice.as_bytes();
-    let bob_state = bob.as_bytes();
-
-    // === Restart alice and bob ===
-    println!("Restart alice and bob");
-    let mut alice_new: RatchetState<PrivateV1Domain> =
-        RatchetState::from_bytes(&alice_state).unwrap();
-    let mut bob_new: RatchetState<PrivateV1Domain> = RatchetState::from_bytes(&bob_state).unwrap();
-
-    // === Alice sends a message ===
-    let (ciphertext, header) = alice_new.encrypt_message(b"Hello Bob!");
-
-    // === Bob receives ===
-    let plaintext = bob_new.decrypt_message(&ciphertext, header);
-    println!(
-        "New Bob received: {}",
-        String::from_utf8_lossy(&plaintext.unwrap())
-    );
-
-    // === Bob replies (triggers DH ratchet) ===
-    let (ciphertext, header) = bob_new.encrypt_message(b"Hi Alice!");
-
-    let plaintext = alice_new.decrypt_message(&ciphertext, header);
-    println!(
-        "New Alice received: {}",
-        String::from_utf8_lossy(&plaintext.unwrap())
-    );
-
-    let (skipped_ciphertext, skipped_header) = bob_new.encrypt_message(b"Hi Alice skipped!");
-    let (resumed_ciphertext, resumed_header) = bob_new.encrypt_message(b"Hi Alice resumed!");
-
-    let plaintext = alice_new.decrypt_message(&resumed_ciphertext, resumed_header);
-    println!(
-        "New Alice received: {}",
-        String::from_utf8_lossy(&plaintext.unwrap())
-    );
-
-    let plaintext = alice_new.decrypt_message(&skipped_ciphertext, skipped_header);
-    println!(
-        "New Alice received: {}",
-        String::from_utf8_lossy(&plaintext.unwrap())
-    );
 }
